+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/components/opaquelinks.js
@@ -34,6 +34,8 @@ function opaqueurl(params) {
 
 function opaqueurl_amend(href, params) {
 
+    params = params || {};
+
     // serialize state into opaque parameter token
     // TODO: make this idempotent by saving the original "href" contents into a "data" attribute
     // if we can reperform the token generation on each click, liveview documents will live forever